Refactor: split auth_routes.py into auth, profile, admin_users
auth_routes.py trimmed to 5 auth routes (login, logout, change-password).
New profile.py handles 16 profile/vacation/absence routes.
New admin_users.py handles 15 admin user-management routes.
Shared helpers (_parse_rates_form, Pydantic schemas) moved to shared.py.
No URL changes.
